+++ src/main/java/com/sparta/realtomatoapp/menu/repository/MenuRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.realtomatoapp.domain.menu.repository;
+package com.sparta.realtomatoapp.menu.repository;
 
-import com.sparta.realtomatoapp.domain.menu.entity.Menu;
+import com.sparta.realtomatoapp.menu.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
