+++ src/main/java/com/sparta/realtomatoapp/order/repository/OrderRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.realtomatoapp.domain.order.repository;
+package com.sparta.realtomatoapp.order.repository;
 
-import com.sparta.realtomatoapp.domain.order.entity.Order;
+import com.sparta.realtomatoapp.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
