+++ src/main/java/com/sparta/realtomatoapp/review/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.realtomatoapp.domain.review.repository;
+package com.sparta.realtomatoapp.review.repository;
 
-import com.sparta.realtomatoapp.domain.review.entity.Review;
+import com.sparta.realtomatoapp.review.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
